+++ src/components/Navbar/NavbarContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+// types
+import type { NavbarContextType } from "./types.js";
+
+export const NavbarContext = createContext<NavbarContextType | undefined>(
+  undefined,
+);