+++ src/components/Notification/types.ts
@@ -1,0 +1,3 @@
+import type { NotificationType } from "lib";
+
+export type NotificationItem = NotificationType & { closing: boolean };