+++ src/providers/index.ts
@@ -1,5 +1,6 @@
 export * from "./NotificationProvider";
 export * from "./ManagerProvider";
+export * from "./AppProviders";
 export * from "./Supabase";
 export * from "./Auth";
 export * from "./ConfigProvider";